refactor: set Cleo bubble domain to match ICON bubble
also with correct lat/long to km conversions
also move input parameters into config file and add gen_supers, gen_gbxs and copy_iconfiles flags to creating inputs

+++ examples/bubble3d/bubble3d.py
@@ -94,6 +94,10 @@ def inputfiles(
     src_config_filename,
     config_filename,
     config_params,
+    gen_config,
+    gen_gbxs,
+    gen_supers,
+    copy_iconfiles,
     isfigures,
 ):
     from cleopy import editconfigfile
@@ -109,16 +113,19 @@ def inputfiles(
         savefigpath.mkdir(exist_ok=True)
 
     ### --- copy src_config_filename into tmp and edit parameters --- ###
-    config_filename.unlink(missing_ok=True)  # delete any existing config
-    shutil.copy(src_config_filename, config_filename)
-    editconfigfile.edit_config_params(config_filename, config_params)
+    if gen_config:
+        config_filename.unlink(missing_ok=True)  # delete any existing config
+        shutil.copy(src_config_filename, config_filename)
+        editconfigfile.edit_config_params(config_filename, config_params)
 
     ### --- delete any existing initial conditions --- ###
     yaml = YAML()
     with open(config_filename, "r") as file:
         config = yaml.load(file)
-    Path(config["inputfiles"]["grid_filename"]).unlink(missing_ok=True)
-    Path(config["initsupers"]["initsupers_filename"]).unlink(missing_ok=True)
+    if gen_gbxs:
+        Path(config["inputfiles"]["grid_filename"]).unlink(missing_ok=True)
+    if gen_supers:
+        Path(config["initsupers"]["initsupers_filename"]).unlink(missing_ok=True)
 
     ### --- input binary files generation --- ###
     # equivalent to ``import bubble3d_inputfiles`` followed by
@@ -126,6 +133,12 @@ def inputfiles(
     inputfiles_script = path2CLEO / "examples" / "bubble3d" / "bubble3d_inputfiles.py"
     python = sys.executable
     cmd = [python, inputfiles_script, path2CLEO, path2build, config_filename]
+    if gen_gbxs:
+        cmd.append("--gen_gbxs")
+    if gen_supers:
+        cmd.append("--gen_supers")
+    if copy_iconfiles:
+        cmd.append("--copy_iconfiles")
     if isfigures[0]:
         cmd.append("--show_figures")
     if isfigures[1]:
@@ -198,6 +211,10 @@ def plot_results(path2CLEO, config_filename, savefigpath):
 # %%
 ### ----------------------------- RUN EXAMPLE ------------------------------ ###
 if args.do_inputfiles:
+    gen_config = True
+    gen_gbxs = True
+    gen_supers = True
+    copy_iconfiles = True
     inputfiles(
         path2CLEO,
         path2build,
@@ -208,6 +225,10 @@ if args.do_inputfiles:
         src_config_filename,
         config_filename,
         config_params,
+        gen_config,
+        gen_gbxs,
+        gen_supers,
+        copy_iconfiles,
         isfigures,
     )
